+++ src/ingest/auth/login/post.ts
@@ -5,15 +5,13 @@
 
 */
 
-// load node utils
+import type { Request, Response } from 'express'
+import type { JwtPayload } from 'jsonwebtoken'
 import { DateTime } from 'luxon'
 
-// load eventhub utils
 import firebase from '../../../utils/firebase'
 import logger from '../../../utils/logger'
 import response from '../../../utils/response'
-import { Request, Response } from 'express'
-import { JwtPayload } from 'jsonwebtoken'
 
 const source = 'ingest/auth/login'
 
@@ -23,21 +21,16 @@ export default async (req: Request, res: Response) => {
 
 		// send email + password for verification, receive login and user object
 		try {
-			login = await firebase.signInWithEmailAndPassword(
-				req.body.email,
-				req.body.password
-			)
+			login = await firebase.signInWithEmailAndPassword(req.body.email, req.body.password)
 		} catch (_error) {
 			return response.badRequest(req, res, { status: 500 })
 		}
 
 		// return ok
-		const expiresIn = Number.parseInt(login.login.expiresIn)
+		const expiresIn = Number.parseInt(login.login.expiresIn, 10)
 		return response.ok(req, res, {
 			expiresIn,
-			expires: DateTime.now()
-				.plus({ seconds: expiresIn })
-				.toISO(),
+			expires: DateTime.now().plus({ seconds: expiresIn }).toISO(),
 
 			token: login.login.idToken,
 			refreshToken: login.login.refreshToken,