+++ src/ingest/pubsub/userTicketRequest.ts
@@ -1,6 +1,6 @@
-import { LoginTicket } from 'google-auth-library'
-import { Request } from 'express'
+import type { Request } from 'express'
+import type { LoginTicket } from 'google-auth-library'
 
 export default interface UserTicketRequest extends Request {
-	user? : LoginTicket
+	user?: LoginTicket
 }