+++ src/ingest/router.ts
@@ -1,14 +1,14 @@
 // import express.router
-// @ts-ignore
+// @ts-expect-error
 import { isIncluded } from '@swrlab/utils/packages/strings'
-import express, { Request, Response } from 'express'
+import express, { type Request, type Response } from 'express'
+import { middleware } from 'express-openapi-validator'
 import { DateTime } from 'luxon'
-import {middleware} from 'express-openapi-validator'
 
 // load swagger UI
 import swaggerUi from 'swagger-ui-express'
-import swaggerDocument from '../../openapi.json'
 import swaggerConfig from '../../config/swagger-ui'
+import swaggerDocument from '../../openapi.json'
 
 // set up router
 const router = express.Router()
@@ -24,16 +24,11 @@ router.use(
 		validateRequests: true,
 		validateResponses: false,
 		ignorePaths: (path: string) =>
-			path.startsWith('/openapi') ||
-			path === '/' ||
-			path === '/health' ||
-			path === '/pubsub',
+			path.startsWith('/openapi') || path === '/' || path === '/health' || path === '/pubsub',
 		formats: {
 			'iso8601-timestamp': {
 				type: 'string',
-				validate: (value) =>
-					isIncluded(value, 'T') &&
-					DateTime.fromISO(value).isValid,
+				validate: (value) => isIncluded(value, 'T') && DateTime.fromISO(value).isValid,
 			},
 		},
 	})
@@ -43,32 +38,23 @@ router.use(
 import response from '../utils/response'
 
 // register swagger endpoints
-router.get('/openapi/openapi.json', (_req: Request, res: Response) =>
-	res.json(swaggerDocument)
-)
-router.get('/openapi/openapi.yaml', (_req: Request, res: Response) =>
-	res.sendFile('openapi.yaml', { root: '.' })
-)
+router.get('/openapi/openapi.json', (_req: Request, res: Response) => res.json(swaggerDocument))
+router.get('/openapi/openapi.yaml', (_req: Request, res: Response) => res.sendFile('openapi.yaml', { root: '.' }))
 router.use('/openapi', swaggerUi.serve, swaggerUi.setup({}, swaggerConfig))
-
-// load auth middleware
-import authVerify from './auth/middleware/verify'
 
 // register API endpoints
 import login from './auth/login/post'
+// load auth middleware
+import authVerify from './auth/middleware/verify'
 import refresh from './auth/refresh/post'
 import reset from './auth/reset/post'
-
-import pubsubAuthVerify from './pubsub/verify'
-
 import events from './events/post'
-
 import pubsub from './pubsub'
-
+import pubsubAuthVerify from './pubsub/verify'
+import subscriptionsDelete from './subscriptions/delete'
+import subscriptionsGet from './subscriptions/get'
 import subscriptionsList from './subscriptions/list'
 import subscriptionsPost from './subscriptions/post'
-import subscriptionsGet from './subscriptions/get'
-import subscriptionsDelete from './subscriptions/delete'
 
 import topics from './topics/list'
 
@@ -84,11 +70,7 @@ router.post('/pubsub/', pubsubAuthVerify, pubsub)
 router.get('/subscriptions/', authVerify, subscriptionsList)
 router.post('/subscriptions/', authVerify, subscriptionsPost)
 router.get('/subscriptions/:subscriptionName', authVerify, subscriptionsGet)
-router.delete(
-	'/subscriptions/:subscriptionName',
-	authVerify,
-	subscriptionsDelete
-)
+router.delete('/subscriptions/:subscriptionName', authVerify, subscriptionsDelete)
 
 router.get('/topics/', authVerify, topics)
 router.get('/topics/:topicName', authVerify, topics)
@@ -99,18 +81,14 @@ router.get(['/', '/health'], (_req: Request, res: Response) => {
 })
 
 // set which error message to return (other may contain private information)
-const allowedErrors = [
-	'Authorization header required',
-	'GET method not allowed',
-]
+const allowedErrors = ['Authorization header required', 'GET method not allowed']
 
 // set openapi error handler
 router.use((err: any, req: Request, res: Response, _next: any) => {
 	// set error message
 	let useOriginalError = false
 	if (allowedErrors.includes(err.message)) useOriginalError = true
-	if (err.message.includes('must have required property'))
-		useOriginalError = true
+	if (err.message.includes('must have required property')) useOriginalError = true
 
 	return response.badRequest(req, res, {
 		message: useOriginalError ? err.message : 'Bad request',