+++ CHANGELOG.md
@@ -5,13 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2025-11-04
+
+- fix: prevent multiple downloads of the ard feed
+- fix: proper publisher lookup
+
 ## [2.0.2] - 2025-10-28
+
 - fix: change dockerfile to use bun instead of node
 
 ## [2.0.1] - 2025-10-28
+
 - chore: redeploy
 
 ## [2.0.0] - 2025-10-16
+
 - refactor: use bun with typescript instead of javascript and node
 - refactor: use bun instead of npm to test
 - fix: retrieve the feed at the application start